model/requirements.txt
brotlipy==0.7.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
cryptography @ file:///C:/b/abs_8ecplyc3n2/croot/cryptography_1677533105000/work
cycler==0.11.0
filelock @ file:///C:/b/abs_c7yrhs9uz2/croot/filelock_1672387617533/work
fonttools==4.39.3
fsspec==2023.5.0
huggingface-hub==0.14.1
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
Jinja2 @ file:///C:/b/abs_7cdis66kl9/croot/jinja2_1666908141852/work
kiwisolver==1.4.4
MarkupSafe @ file:///C:/ci/markupsafe_1654508036328/work
matplotlib==3.7.1
mkl-fft==1.3.6
mkl-random @ file:///C:/Users/dev-admin/mkl/mkl_random_1682977971003/work
mkl-service==2.4.0
mpmath==1.2.1
networkx @ file:///C:/b/abs_b935xy_9g6/croot/networkx_1678964342510/work
numpy==1.23.5
packaging==23.1
Pillow==9.4.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///C:/b/abs_552w85x1jz/croot/pyopenssl_1677607703691/work
pyparsing==3.0.9
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
python-dateutil==2.8.2
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
six==1.16.0
sympy @ file:///C:/b/abs_95fbf1z7n6/croot/sympy_1668202411612/work
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
transformers==4.28.1
typing_extensions @ file:///C:/b/abs_a1bb332wcs/croot/typing_extensions_1681939523095/work
urllib3==2.0.2
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work

